split domain code in separate files
