+++ nextjs/src/domain/error.ts
@@ -1,0 +1,3 @@
+export class NotFoundError extends Error {}
+export class UnauthorizedError extends Error {}
+export class ForbiddenError extends Error {}
